* commit 'a99517990da67b4c1f23777246d97bba01b85612':
  Optimze time sharing stock view;

+++ src/com/dkhs/portfolio/bean/StockQuotesBean.java
@@ -45,9 +45,9 @@ public class StockQuotesBean {
     // 换手率
     private float turnover_rate;
     // 流通市值 单位：元
-    private float market_capital;
+    private long market_capital;
     // 总市值 单位：元
-    private float total_capital;
+    private long total_capital;
     // 动态市盈率
     private float pe_ttm;
     // 市净率
@@ -252,19 +252,19 @@ public class StockQuotesBean {
         this.turnover_rate = turnover_rate;
     }
 
-    public float getMarket_capital() {
+    public long getMarket_capital() {
         return market_capital;
     }
 
-    public void setMarket_capital(float market_capital) {
+    public void setMarket_capital(long market_capital) {
         this.market_capital = market_capital;
     }
 
-    public float getTotal_capital() {
+    public long getTotal_capital() {
         return total_capital;
     }
 
-    public void setTotal_capital(float total_capital) {
+    public void setTotal_capital(long total_capital) {
         this.total_capital = total_capital;
     }
